// src/app/models/yals.model.ts
export class YalsRequest {
    street: String;
    num_ext: String;
    zip: String;
    colonia: String;
    municipio: String;
    estado: String;
    ciudad: String;
    id_tipo_propiedad: Number;
    recamaras: Number;
    banos: Number;
    medios_banos: Number;
    estacionamientos: Number;
    area_construida: Number;
    superficie_terreno: Number;
    edad: Number;
    amenities: String[];
    email: String;
    api_key: String;
    latitud: Number;
    longitud: Number;
}